addons/ldap-passwd-expiry/distrib/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
pyasn1-modules==0.2.6     # via python-ldap
pyasn1==0.4.6             # via pyasn1-modules, python-ldap
python-ldap==3.2.0
slackclient==2.5.0
requests==2.22.0
